refactor: 마이페이지 프로필 컨테이너 내 사용자 관련 로직 서비스 함수로 분리
